flask-api/requirements.txt
Flask==2.3.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1